+++ castle/cms/cron/_twitter_monitor.py
@@ -6,6 +6,12 @@ from datetime import datetime
 import transaction
 from AccessControl.SecurityManagement import newSecurityManager
 from BTrees.OOBTree import OOBTree
+from castle.cms.indexing.hps import index_in_es
+from castle.cms.services import twitter
+from castle.cms.social import COUNT_ANNOTATION_KEY
+from castle.cms.utils import retriable
+from castle.cms.services.twitter import get_auth
+from datetime import datetime
 from plone.app.redirector.interfaces import IRedirectionStorage
 from plone.registry.interfaces import IRegistry
 from Products.CMFPlone.interfaces.siteroot import IPloneSiteRoot
@@ -14,12 +20,6 @@ from tendo import singleton
 from zope.annotation.interfaces import IAnnotations
 from zope.component import getUtility
 from zope.component.hooks import setSite
-
-from castle.cms.services import twitter
-from castle.cms.services.twitter import get_auth
-from castle.cms.social import COUNT_ANNOTATION_KEY
-from castle.cms.utils import index_in_es
-from castle.cms.utils import retriable
 
 
 def record_tweet_stats(site, ob, tweet):