+++ castle/cms/services/twitter.py
@@ -352,9 +352,9 @@ class Stream(object):
         if resp.raw.closed:
             self.on_closed(resp)
 
-    def _start(self, async):
+    def _start(self, is_async):
         self.running = True
-        if async:
+        if is_async:
             self._thread = Thread(target=self._run)
             self._thread.start()
         else:
@@ -370,7 +370,7 @@ class Stream(object):
                    replies=None,
                    track=None,
                    locations=None,
-                   async=False,
+                   is_async=False,
                    encoding='utf8'):
         self.session.params = {'delimited': 'length'}
         if self.running:
@@ -387,38 +387,38 @@ class Stream(object):
             if len(locations) % 4 != 0:
                 raise Exception("Wrong number of locations points, "
                                 "it has to be a multiple of 4")
-            self.session.params['locations'] = ','.join(['%.2f' % l for l in locations])
+            self.session.params['locations'] = ','.join(['%.2f' % l for l in locations])  # noqa: E741
         if track:
             self.session.params['track'] = u','.join(track).encode(encoding)
 
-        self._start(async)
+        self._start(is_async)
 
-    def firehose(self, count=None, async=False):
+    def firehose(self, count=None, is_async=False):
         self.session.params = {'delimited': 'length'}
         if self.running:
             raise Exception('Stream object already connected!')
         self.url = '/%s/statuses/firehose.json' % STREAM_VERSION
         if count:
             self.url += '&count=%s' % count
-        self._start(async)
+        self._start(is_async)
 
-    def retweet(self, async=False):
+    def retweet(self, is_async=False):
         self.session.params = {'delimited': 'length'}
         if self.running:
             raise Exception('Stream object already connected!')
         self.url = '/%s/statuses/retweet.json' % STREAM_VERSION
-        self._start(async)
+        self._start(is_async)
 
-    def sample(self, async=False, languages=None):
+    def sample(self, is_async=False, languages=None):
         self.session.params = {'delimited': 'length'}
         if self.running:
             raise Exception('Stream object already connected!')
         self.url = '/%s/statuses/sample.json' % STREAM_VERSION
         if languages:
             self.session.params['language'] = ','.join(map(str, languages))
-        self._start(async)
+        self._start(is_async)
 
-    def filter(self, follow=None, track=None, async=False, locations=None,
+    def filter(self, follow=None, track=None, is_async=False, locations=None,
                stall_warnings=False, languages=None, encoding='utf8', filter_level=None):
         self.body = {}
         self.session.headers['Content-type'] = "application/x-www-form-urlencoded"
@@ -433,7 +433,7 @@ class Stream(object):
             if len(locations) % 4 != 0:
                 raise Exception("Wrong number of locations points, "
                                 "it has to be a multiple of 4")
-            self.body['locations'] = u','.join(['%.4f' % l for l in locations])
+            self.body['locations'] = u','.join(['%.4f' % l for l in locations])  # noqa: E741
         if stall_warnings:
             self.body['stall_warnings'] = stall_warnings
         if languages:
@@ -442,10 +442,10 @@ class Stream(object):
             self.body['filter_level'] = six.text_type(filter_level, encoding)
         self.session.params = {'delimited': 'length'}
         self.host = 'stream.twitter.com'
-        self._start(async)
+        self._start(is_async)
 
     def sitestream(self, follow, stall_warnings=False,
-                   with_='user', replies=False, async=False):
+                   with_='user', replies=False, is_async=False):
         self.body = {}
         if self.running:
             raise Exception('Stream object already connected!')
@@ -458,7 +458,7 @@ class Stream(object):
             self.body['with'] = with_
         if replies:
             self.body['replies'] = replies
-        self._start(async)
+        self._start(is_async)
 
     def disconnect(self):
         if self.running is False: