+++ castle/cms/tiles/gallery.py
@@ -56,7 +56,8 @@ class IGalleryTileSchema(IImagesTileSchema):
         title=u'Show Thumbnails',
         description=(
             u'Display thumbnail images below main gallery image. '
-            u'This option is for advanced display types in certain add-ons.Usually, this option has no affect.'
+            u'This option is for advanced display types in certain \
+                add-ons.Usually, this option has no affect.'
         ),
         missing_value=True,
         required=False,
@@ -66,7 +67,8 @@ class IGalleryTileSchema(IImagesTileSchema):
         title=u'Show Image Title',
         description=(
             u'Display Image Title directly below the main gallery image. '
-            u'This option is for advanced display types in certain add-ons. Usually, this option has no affect.'
+            u'This option is for advanced display types in certain add-ons. \
+                Usually, this option has no affect.'
         ),
         missing_value=True,
         required=False,
@@ -75,28 +77,33 @@ class IGalleryTileSchema(IImagesTileSchema):
     show_image_captions = schema.Bool(
         title=u'Show Image Captions',
         description=(
-            u'Display a small caption (the image\'s summary field) below either the image title (if visible) or the main image. '
-            u'This option is for advanced display types in certain add-ons. Usually, this option has no affect.'
+            u'Display a small caption (the image\'s summary field) below either \
+                the image title (if visible) or the main image. '
+            u'This option is for advanced display types in certain add-ons. \
+                Usually, this option has no affect.'
         ),
         missing_value=True,
         required=False,
     )
 
-    show_image_counter =  schema.Bool(
+    show_image_counter = schema.Bool(
         title=u'Show Image Counter',
         description=(
-            u'Display a small indicator above the main image indicating which image is being viewed (ex: "Image 15 of 22"). '
-            u'This option is for advanced display types in certain add-ons. Usually, this option has no affect.'
+            u'Display a small indicator above the main image indicating which \
+                image is being viewed (ex: "Image 15 of 22"). '
+            u'This option is for advanced display types in certain add-ons. \
+                Usually, this option has no affect.'
         ),
         missing_value=False,
         required=False,
     )
 
-    show_link_to_original_image =  schema.Bool(
+    show_link_to_original_image = schema.Bool(
         title=u'Show Link to Original Image',
         description=(
             u'When the gallery image is highlighted, show a link to the original image. '
-            u'This option is for advanced display types in certain add-ons. Usually, this option has no affect.'
+            u'This option is for advanced display types in certain add-ons. Usually, \
+                this option has no affect.'
         ),
         missing_value=False,
         required=False,
@@ -112,4 +119,3 @@ class IGalleryTileSchema(IImagesTileSchema):
     #     missing_value=False,
     #     required=False,
     # )
-