+++ castle/cms/upgrades/__init__.py
@@ -1,10 +1,10 @@
 import importlib
 
-import plone.api as api
-from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.resources.browser.cook import cookWhenChangingSettings
-
 from castle.cms.interfaces import IAPISettings
+from castle.cms.interfaces import ISecuritySchema
+from plone import api
+from Products.CMFPlone.resources.browser.cook import cookWhenChangingSettings
+from Products.CMFCore.utils import getToolByName
 
 
 PROFILE_ID = 'profile-castle.cms:default'
@@ -12,6 +12,14 @@ PROFILE_ID = 'profile-castle.cms:default'
 
 def cook_js_resources(context, logger=None):
     cookWhenChangingSettings(api.portal.get())
+
+
+def re_register_profile(interface, prefix):
+    registry = api.portal.get_tool('portal_registry')
+    registry.registerInterface(
+        interface,
+        prefix=prefix,
+    )
 
 
 def custom_upgrade_factory(profile_id):
@@ -38,11 +46,18 @@ upgrade_3000 = default_upgrade_factory('3000')
 upgrade_3001 = default_upgrade_factory('3001')
 upgrade_3003 = default_upgrade_factory('3003')
 upgrade_3004 = default_upgrade_factory('3004')
+upgrade_3005 = default_upgrade_factory('3005')
 
 
-def upgrade_3005(site, logger=None):
-    registry = api.portal.get_tool('portal_registry')
-    registry.registerInterface(
-        IAPISettings,
-        prefix='castle',
-    )
+def upgrade_3006_cc(site, logger=None):
+    # for explicit enable/disable of country code checking
+    re_register_profile(ISecuritySchema, 'castle')
+
+
+def upgrade_3006_ga4(site, logger=None):
+    # for ga4 migration
+    re_register_profile(IAPISettings, 'castle')
+    old_ga_id = api.portal.get_registry_record('castle.google_analytics_id')
+    if old_ga_id:
+        api.portal.set_registry_record('castle.universal_analytics_id', old_ga_id)
+        api.portal.set_registry_record('castle.google_analytics_id', u'')